Build the app so that it can export
I want to export the run() function
so that I could test it if I wanted (it injects the document and console, could inject more). Somehow I missed that it completely didn't work with the build.

By building as an esm, the export of the entrypoint is maintained and we keep the export of run()