Record that someone signs up
Lots of things moved around here,
but the basic idea is that firestore
contains the meetup and the list
of people who have signed up.

When someone clicks on the "Sign Up" button it will either
record in firestore that they are signed up
or it will take them through the verification flow. It all depends on if they are signed in (email verified) or not.

The services class is getting a bit big
and gnarly now. Should start
getting it tested and split up.